port to typescript
