+++ src/impl/네이버뉴스.ts
@@ -1,0 +1,30 @@
+import { clearStyles } from '../util';
+import { Article } from 'index';
+
+export function parse(): Article {
+    return {
+        title: document.querySelector('#articleTitle').innerText,
+        content: (() => {
+            const content = document.querySelector('#articleBodyContents');
+            const iframes = content.querySelectorAll('.vod_area iframe[_src]');
+            if (iframes.length > 0) {
+                iframes.forEach((v) => {
+                    v.setAttribute('src', v.getAttribute('_src'));
+                });
+            }
+            return clearStyles(content).innerHTML;
+        })(),
+        timestamp: {
+            created: (() => {
+                let created = document.querySelector('.article_info .sponsor .t11').innerText;
+                return new Date(created.replace(' ', 'T') + '+09:00'); // ISO 8601
+            })(),
+            lastModified: undefined
+        },
+        cleanup: () => {
+            document.querySelectorAll('#tooltipLayer_english, .u_cbox_layer_wrap').forEach(function (v) {
+                v.remove();
+            });
+        }
+    };
+}
